--- src/main/java/com/dinosaur/foodbowl/domain/auth/application/CookieService.java
@@ -1,19 +1,0 @@
-package com.dinosaur.foodbowl.domain.auth.application;
-
-import jakarta.servlet.http.Cookie;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class CookieService {
-
-  public Cookie generateCookie(String name, String value, int expire) {
-    Cookie cookie = new Cookie(name, value);
-    cookie.setHttpOnly(true);
-    cookie.setMaxAge(expire);
-    return cookie;
-  }
-}
