+++ src/test/java/com/dinosaur/foodbowl/IntegrationTest.java
@@ -7,7 +7,7 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 
 import com.dinosaur.foodbowl.domain.auth.application.AuthService;
-import com.dinosaur.foodbowl.domain.auth.application.CookieService;
+import com.dinosaur.foodbowl.global.util.CookieUtils;
 import com.dinosaur.foodbowl.domain.auth.application.TokenService;
 import com.dinosaur.foodbowl.domain.category.dao.CategoryRepository;
 import com.dinosaur.foodbowl.domain.follow.application.FollowService;
@@ -89,7 +89,7 @@ public class IntegrationTest {
   protected TokenService tokenService;
 
   @SpyBean
-  protected CookieService cookieService;
+  protected CookieUtils cookieUtils;
 
   @SpyBean
   protected DeleteAccountService deleteAccountService;
